Fix Add and Sub implementations for Scalar
The contract of `Scalar` is that for any `x` in `Scalar`,
`x mod p === x mod q`, where `p` and `q` are the orders of secp256k1
and ed25519.

To ensure that this is always the case when doing arithmetic
operations on Scalar, we must operate modulo the smaller of the two
curve orders.

Since ed25519 has a smaller curve order than secp256k1,
we operate modulo ed25519.