Limit Scalars to 252 bits
Scalars were already being generated correctly, given that we were
using the scalar generation provided by the curve with the smaller
curve order out of the two i.e. ed25519.

This generated scalars to up to `2^252 +
0x14def9dea2f79cd65812631a5cf5d3ed`, which required a maximum of 253
bits. To be able to define a maximum bit length for the `Scalar` type
we would have to go to 252 bits, since some 253-bit scalars would
require reduction modulo ed25519 curve order.

The advantage of limiting the number of bits of `Scalar` is that the
proof elements can also be limited to arrays of 252 elements, meaning
that we can produce 4 fewer tuples for those bits that are always
zero. This should improve performance and reduce the likelihood of
mistakes.